requirements.txt
tensorflow==2.13.0
fastai==2.7.12
torch==2.0.1
torchvision==0.15.2
mlflow==2.8.1
psycopg2-binary
boto3
minio
tqdm
requests
pillow
scikit-learn
fastapi
uvicorn
python-multipart
pandas==1.5.3
numpy==1.24.3
matplotlib
fastcore==1.5.29
pymysql
python-dotenv
gradio
Pillow